# src/documentation/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Nutriguide API",
    "description": "Documentação dos endpoints do backend do projeto Nutriguide, desenvolvido por Henrique Torres (https://github.com/henriqtorresl)",
    "version": "1.0.0"
  },
  "host": "https://nutriguide-api.vercel.app",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/login": {
      "post": {
        "tags": [
          "Auth Resource"
        ],
        "description": "Faz o login na aplicação e gera um token jwt",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/cadastro": {
      "post": {
        "tags": [
          "Auth Resource"
        ],
        "description": "Faz o cadastro de um nutricionista na aplicação e gera um token jwt",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user/nome": {
      "get": {
        "tags": [
          "Auth Resource"
        ],
        "description": "Trás os dados pessoais do usuário a partir do seu nome",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "nome",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  }
}